requirements.txt
Flask
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
pyjwt